Create separate test file for ASR models and create separate ci group for remote code models
