+++ .github/workflows/test_openvino.yml
@@ -70,11 +70,6 @@ jobs:
         run: |
           uv pip install transformers==${{ matrix.transformers-version }}
 
-      - if: ${{ matrix.transformers-version == '4.57.6' }}
-        name: Install qwen-asr funasr packages
-        run: |
-          uv pip install qwen-asr funasr
-
       - if: ${{ matrix.transformers-version == '4.45.0' }}
         name: Install specific dependencies and versions required for older transformers
         run: |
@@ -92,7 +87,7 @@ jobs:
 
       - name: Test with Pytest
         run: |
-          pytest tests/openvino/${{ matrix.test-pattern }} --durations=0
+          pytest tests/openvino/${{ matrix.test-pattern }} -k "not (qwen3_asr or fun_asr)" --durations=0
 
       - if: ${{ matrix.test-pattern == '*modeling*' }}
         name: Install Nightly OpenVINO
