+++ .github/workflows/test_openvino_remote_code.yml
@@ -1,0 +1,77 @@
+name: OpenVINO - Remote Code Models Validation
+
+on:
+  workflow_dispatch:
+  push:
+    branches:
+      - main
+      - v*-release
+  pull_request:
+    branches:
+      - main
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
+  cancel-in-progress: true
+
+permissions:
+  contents: read
+
+env:
+  UV_TORCH_BACKEND: cpu
+  UV_SYSTEM_PYTHON: true
+  TRANSFORMERS_IS_CI: true
+  HF_TOKEN: ${{ secrets.HF_HUB_READ_TOKEN }}
+
+jobs:
+  build:
+    strategy:
+      fail-fast: false
+      matrix:
+        test-pattern:
+          [
+            "*export*",
+            "*asr*",
+            "*quantization*",
+          ]
+
+    runs-on: ubuntu-22.04
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v7.0.0
+
+      - name: Setup Python
+        uses: actions/setup-python@v6
+        with:
+          python-version: "3.10"
+
+      - name: Install dependencies
+        run: |
+          pip install --upgrade pip uv
+          uv pip install .[tests] librosa
+
+      - name: Restore HF models cache
+        id: hf-cache-restore
+        if: env.HF_TOKEN == ''
+        uses: actions/cache/restore@v6
+        with:
+          path: ~/.cache/huggingface/hub
+          key: hf-models
+
+      - name: Install transformers
+        run: |
+          uv pip install transformers==4.57.6
+
+      - name: Install qwen-asr funasr packages
+        run: |
+          uv pip install qwen-asr funasr
+
+      - name: Login with fork PRs CI token
+        if: ${{ env.HF_TOKEN == '' }}
+        run: |
+          python tests/scripts/login_with_ci_token.py
+
+      - name: Test Qwen3-ASR and FunASR
+        run: |
+          pytest tests/openvino/${{ matrix.test-pattern }} -k "qwen3_asr or fun_asr" --durations=0
